pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "manel"
version = "0.1.0"
description = "AI-powered manga-to-panel reader for Kindle"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.4",
    "torchvision>=0.19",
    "transformers>=4.45",
    "accelerate>=0.34",
    "sam2>=1.0",
    "surya-ocr>=0.8",
    "pypdf>=5.0",
    "pillow>=10.4",
    "gradio>=5.0",
    "ebooklib>=0.18",
    "lxml>=5.3",
    "pydantic>=2.9",
    "click>=8.1",
    "tqdm>=4.66",
    "numpy>=2.0",
    "scikit-image>=0.24",
    "ultralytics>=8.4",
    "flet>=0.84",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[project.scripts]
manel = "manel.cli:main"

[tool.setuptools.packages.find]
include = ["manel*"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
